requirements.txt
bidict==0.21.2
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-engineio==4.0.0
python-socketio==5.0.4
Werkzeug==1.0.1
